keep a reference to the semanticdb reporter
Keep a reference to the reporter we're overriding. Other compiler
plugins (silencer compiler plugin) can change the reporter and we're
not guaranteed that it will be a StoreReporter. We can end up in a
situation where the reporter is replaced and the messages aren't
passed to the underlying reporter or ignored.