Introduce `ExportGiven`, remove `Export.given: Boolean`.
Previously, the `scala.meta.Export` tree had a `given: Boolean` field.
This goes against a design principle in the Scalameta AST, which tries
to represent as much as possible with syntax. There's no other AST node that
has a boolean field to distinguish between two cases. This commit
introduces a new `ExportGiven` class to represent `export given a.b`
syntax.

This is a binary breaking change so the next release should strictly
speaking be Scalameta v5. However, I think it's fair to change the
compatibility policy to state that Dotty-specific tree nodes are
experimental for time being and subject to breaking changes. This policy
buys us time to refine the Scala 3 AST nodes so that they are well
designed and consistent with the rest of the Scalameta AST nodes.
Normal Scalameta users should not be affected by this change as long as
they don't use Scalameta to parse Scala 3 programs.