Fix givens to accept `with` keyword for refinements
In the previous PR that touched the changes in givens I forgot to include `with`, which I am fixing in this PR. I also removed any test cases that were ignored, but don't have an easy way for the parser to show an error (also Dotty parser mostly works in those case also).

Aside from adding `with` I also made sure that multiple inheritance in givens works as expected and using indentation is possible.