Make sure children contains binary compat fields
Binary compat fields were added to allow for the older AST nodes to be forward binary compatible.

Previously, when invoking `children` method on trees that contained binaryCompat fields, those fields would not exist in the resulting list. Now, they are contained and can be used to traverse the tree.

It's important to note that traverse will still not working with those additional fields, but in most places (for example scalafmt) the `children` method is being used. We can try to fix it in the future if need arises.