Parser: rework significant indentation processing.
It is mostly about rework. In addition, it fixes `multiline-extension` parsing.

The logic from `loop` method that removes whitespaces and produces `Intent`/`Outdent` is moved on `TokenIterator` level.
It observes tokens on demand that adds an ability to put indentation on the parser level using `observeIndented` method.
Dotty parser uses a similar approach.

So, instead of trying to predict _if this `LF` that comes after `:` should add Indent or not_ now it's handled at the moment of parsing template. The same for extension methods.