Add testkit/test to CI and ignore failing test
Traversal doesn't work correctly for @binaryCompatFields currently. This will be an issue for anyone using it for Scala 3 sources, but should not be a problem otherwise. To fix it we need to make traversal work  with those fields or bump the version to 5.0.0 and turn all of those fields to real ones.

Fixes https://github.com/scalameta/scalameta/issues/2246