[Scala 3] Fix given alias with complex types
Complex types such as `A & B` need to be wrapped in parenthesis when used as a type for given alias. Previously, we would try to parse parens as using parameters, which is not always the case. Now we also check for the `using` keyword, which allows us to parse complex types in given aliases type declarations.