[Scala 3] Allow return statement to contain indented block
Previously, return could not start an indentation, but starting with https://github.com/lampepfl/dotty/pull/11752 it can. This fixes scalameta to parse accordingly.

Additionally, this change also created some issues with empty `return` statements in pattern matches. We would incorrectly think that the next statement is indented, because we never saved case block indentation. I added a new CaseRegion in this PR, so that we can remember the indentation. This is only added in `match` with braces as indented match clauses work correctly here.