Trees: add an option to get original input syntax
Currently, the syntax is requested with an implicit dialect, and most of
the time the dialect isn't set explicitly but uses whatever happened to
be provided by default.

For this type of case, let's provide a way to obtain the original syntax
as it was used when parsing the code.