[Scala 3] Add the rest of possible inline matched expressions
In the previous PR, I haven't added all possible cases, which inline can be followed by with a match. I went through all ExprIntro tokens and checked each one, so we should be good now. We could use `ExprIntro` classifier in theory, but it seems around half of the tokens defined there would be invalid after inline.

This issue came up when I was looking through the Dotty test cases and it seems we should try to parse as many as possible there, however it seems we do have some bugs still and I don't want to fix it all in one PR.