Fix UnreachableError when printing a pattern match bind to a case object
Fixes #2447

`{ case x => Y => () }` parses differently than `{ case x @ y => () }`. I'm assuming this is on purpose and for a very good reason.

This PR changes the `thisLocationAlsoAcceptsPatVars` to return true for a `Pat.Bind`. I'm _assuming_ this is correct, as in a pattern match a `Pat.Var` can also take the place of a `Pat.Bind` (like in `{ case x => () }`).