chore: split semanticdb and parser tests to migrate to sbt 1.7.1
Due to `++ command updates` in https://github.com/sbt/sbt/releases/tag/v1.7.0 ++ is much more strick and requires the Scala version to be available in cross versions. This causes a bunch of problems with modules that we don't want to publish for all Scala versions, however we really want to test all the Scala versions only for semanticdb. Because of that I decided to split them into a separate project.