bugfix: Manually skip newlines after soft keywords
Previously, we would always skip newlines between `open` identifier and the next token if that could start a statement. This led to lossing info about newlines even if the open modifier was not used. Now, we skip newlines after we know a soft keyword is used.

Fixes https://github.com/scalameta/scalameta/issues/2786