Emit local symbol occurrence for structural type access
Previously, semanticdb-scalac never emitted occurrences for invocations
of members on structural types. The motivation for this behavior was
that there was no reasonable symbol we could use since there no
official defintion for a structural type, the same structural type
can be re-declared in multiple places. This commit changes the
behavior to emit local occurrences when the structural type has been
defined in the same file as the reference occurrence. We still don't
emit reference occurrences when the structural type is referenced
from a separate source file.