ScalametaParser: constrain match type pattern
Currently, unlike pattern matching for Terms which follows a different
parsing logic and produces Pat trees, pattern matching for Types uses
exactly the same logic as regular Types.

This occasionally leads to incorrect parsing, and here we try to address
the specific problem of parsing a wildcards and anonymous params which
is different in a type pattern match context.