ast: allow replacing old fields with new ones
As an alternative to existing `newFields` which are appended at the end
of the list of existing params, let's have the ability to remove a field
as long as it's replaced with another.

Current implementation supports cases when the new field value can be
determined from the old one via an implicit conversion, similar to how
quasiquotes define implicit Lift/Unlift conversions.

This change is backwards-binary-compatible and hence will not trigger
MIMA-verification alerts.