bugfix: Don't allow * as type lambda by default
This currently needs to be enabled with a compiler flag, so I think it makes sense not to have it on as default. In Metals we can properly detect the flag, while in scalafmt people will most likely need to set it themselves.

Related to https://github.com/scalameta/metals/issues/5573