improvement: Remove macro usage from Scalameta parser
I was trying to get the parsers code to compile with Scala 3, but the classifier macros were a bit of a problem. Rewriting them would require other modules to also cross compile, while removing makes it possible to just use other modules as Scala 2.13 ones.

The changes are not terrible, so I feel like the macros weren't really neccessary, while obviously useful.