ScannerTokens: make `isLeadingInfix` compliant
The implementation was incorrect, didn't check the case when the infix
op was on a line by itself.

However, the fix is incomplete as it doesn't handle the very rare case
when the infix expression is at the file level and not nested within a
block of any kind. This will be fixed in the next commit.

Also, add a correlated fix in `getOutdentIfNeeded` and introduce a new
token, InfixLF, so that we don't need to repeat the logic in the parser.