ast: implicit dialect in apply(), store in origin
That way, all our trees will have a proper origin containing a
dialect.

Move previous versions of apply, those without dialect, to a new trait,
and extend companion object from that trait. That way, MIMA is satisfied
and compiler doesn't complain of similar methods matching invocation.

One complexity arises for trees which are defined at the package level
rather than within an object. In that case, we can't use that new trait
as the macro is not allowed to emit extraneous types at top level.

In that case, we will not pass the implicit dialect but will determine it
from one of the trees which were passed as arguments to apply().