ScannerTokens: improve handling of `finally`
Since we allow `finally` to be at the same indent as the `try` body,
we could end up with `finally` terminating both the previous `catch`
as well as the outer `try`.

Instead, let's introduce a new `RegionTry` and ensure that this region
is not removed before the `finally` keyword, so that we won't end up
with an exposed outer `RegionTry`.