Transversers: skip Quasi before other trees
Since every Quasi tree in reality derives from one of the other types,
and the declaration of a Quasi tree follows that of its parent, the
case for the parent happens before the case for the quasi, and hence
the quasi code is unreachable.

Since the intention was likely to exclude Quasi trees altogether, let's
skip them at the beginning rather than the end.