TextDocumentOps: don't index empty names
For instance, we might substitute a `Type.Name("Unit")` when the type
is missing in a method declaration, but its input position is empty as
it isn't actually present in the itput, which might lead to errors due
to hash collisions with a parameterless-method name.

Prior to v4.9.4, it "worked" since the position of the empty name was
actually around the newline token rather than the preceding method-name
token.