ScalametaParser: less backtrack on ctrl cond, body
If a scala3 if/while actually uses scala2 syntax, we might have parsed
part of the body while looking for `then` or `do`, which leads to some
backtracking.

Specifically, for nested `if` clauses with their `then` part in explicit
braces, each will be parsed exactly twice, once as a continuation of the
condition expression and second time as the body, thus leading to an
exponential time complexity.

Let's instead reduce backtracking by processing what might be both an
argument to a condition expression or an initial part of the body,
such as a brace expression, and then using it either for condition or
for body.