LegacyScanner: no quotes in interpolation tokens
Currently, legacy string literal (STRINGLIT) tokens are emitted together
with the quotes around them, as those are also included in the resulting
Token.Constant.String.

On the other hand, when it comes to interpolation strings, the quotes
are instead included in the Interpolation.Start and .End tokens, but
there's an inconsistency in implementation.

The opening quotes are derived from the gap between INTERPOLATIONID and
STRINGPART, the first two legacy tokens of an interpolation. The final
legacy token of an interpolation is a STRINGLIT, which, however, does
include the closing quotes, which later need to be adjusted for using a
crude approximation which doesn't take into account possible escaping
or unicode literals used to represent those quotes.

Let's change the implementation to exclude the closing quotes from the
final interpolation STRINGLIT.