bugfix: Use earliest Scala version for scalameta, parsers and trees proejcts
If we use latest by default then the semanticdb parts will also depend on newer Scala versions, which is breaking for anything that might be using it as dependency.

The other option here is to change their suffix to full and publish separate for each version, but I would rather avoid it as it would mean more artifacts to publish and a change for any dependency.