bugfix: Set earlier versions as only cross versions for scalameta projects
Turns out, the previous approach didn't work and we still published with the newest Scala version. I changed it to only use the earliest versions, so that it's not even possible to set newer versions. I also changed sharedSettings to not include cross settings by default as they should be defined per project (or with fullCrossSettings).