Input: add WithTokenizerOptions wrapper
Since our public code signatures take Input and Dialect, the only way
to provide TokenizerOptions without breaking MIMA compatibility is by
using either of those types.

Neither of them in reality is perfect for this task, but since Dialect
is a final class and adding tokenizer options as a field to it seems
a bit illogical, let's implement it as a wrapper of the Input type.