TextDocumentOps: traverse function properly
In a recent version of compiler, it uses incorrect position values for
anonymous functions which leads to incorrect symbol matching for that
function instead of its body.

Hence, make sure we don't allow the function to get to the fallback case
in the match and explicitly traverse parameters and body separately.