TreeSyntax: look for brace, not just Term.Block
In the logic deciding whether to insert a newline between two stats,
checking if the next stat is a block is insufficient, as it could be
an expression (select or infix chain) which *starts* with a block.

Instead, let's turn it into a Show.Result and check if it starts with
a brace.