Fix issue with not recognising dialects in post-3.4.0 versions of Scala
It looks like implicit vals in scala 3.3.x are unpickled into a methods,
and in later version they are unpickled into vals. To confirm that the
imported dialect can be found in the proper package, we now just check
whether it's a declaration there, regardless of it being a method or
a val.