chore: Update to 2.13.17 (#4316)
I had to change the behaviour in extractors, because of change in positions for Scala 2.13.17.

I added scala version to make sure behaviour stays the same for older versions.

Some tests also needed to be updated and SemanticdbSuite started showing errors in the reporter, but it didn't break the actual results, so ignored it for now.