ReificationMacros: fall back to partial origin
Currently, we use Origin.Parsed if quasiquote source has no holes (or
no interpolations), and DialectOnly otherwise.

Instead of DialectOnly, let's expose some partial information about the
origin (that is, everything but text associated with the position since
that would refer to unexpanded tokens) which would allow us to compare
token positions if not their content.