ast: fix mima problems by restoring old apply
Since we have added new private fields, the "private field" portion of
the apply methods has changed. Instead, we should preserve the old apply
signatures and add the new ones.

For that, let's mark each private field with the version after which it
was added (and the original ones will refer to v0.0.0). And for each of
these versions, we'll generate a variant of apply, with an appropriate
signature.