Tokenizer: add {Int,Float}XL tokens and trees
In tokenizer, extract fractional-value tokens as Constant.FloatXL unless
they have an explicit suffix identifying it as Float or Double, or if
they are within Double range; at the same time, integer tokens that do
not have a suffix will continue to be extracted as Constant.Int if they
are within 32-bit, and as Constant.IntXL otherwise.

Add a new AnyDecimal "numeric" type which refers to arbitrary-precision
decimal values.